+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## [Unreleased]
+### Added
+- Add `AccountSecurity` endpoint and resource
+
 ## [1.3.0] - 2021-09-03
 ### Added
 - Add `EnvelopeTemplateDocument.positions` endpoint